pyproject.toml
[project]
name = "langchain-conversationrelay"
version = "0.1.0"
description = "Voice-capable AI agent using LangChain 1.x and Twilio ConversationRelay with persistent memory and dual text/voice interfaces"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.124.0",
    "langchain>=1.0,<2.0",
    "langchain-anthropic>=1.2.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.1",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "twilio>=9.8.8",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
